+++ packages/_private/vanilla-demo/src/AllComponentParts/Select.ts
@@ -49,7 +49,7 @@ export const Select = (regressions: Array<() => void> = []) => {
     ></syn-select>
 
     <syn-select
-      data-testid="select-851-multiple"
+      data-testid="select-847-multiple"
       help-text="Normal value binding and async options"
       label="Multiple with async options"
       multiple