+++ packages/components/src/components/side-nav/side-nav.component.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/unbound-method */
 import { classMap } from 'lit/directives/class-map.js';
-import type { CSSResultGroup } from 'lit';
+import type { CSSResultGroup, PropertyValues } from 'lit';
 import { html } from 'lit/static-html.js';
 import { property, query, state } from 'lit/decorators.js';
 import { HasSlotController } from '../../internal/slot.js';
@@ -9,12 +9,14 @@ import componentStyles from '../../styles/component.styles.js';
 import styles from './side-nav.styles.js';
 import SynDrawer from '../drawer/drawer.component.js';
 import SynDivider from '../divider/divider.component.js';
+import SynIcon from '../icon/icon.component.js';
+import SynNavItem from '../nav-item/nav-item.component.js';
 import { waitForEvent } from '../../internal/event.js';
 import { watch } from '../../internal/watch.js';
 import { getAnimation, setAnimation, setDefaultAnimation } from '../../utilities/animation-registry.js';
 import { LocalizeController } from '../../utilities/localize.js';
 import { unlockBodyScrolling } from '../../internal/scroll.js';
-import { blurActiveElement } from '../../internal/closeActiveElement.js';
+import { enableDefaultSettings } from '../../utilities/defaultSettings/decorator.js';
 
 /**
  * @summary The <syn-side-nav /> element contains secondary navigation and fits below the header.
@@ -32,10 +34,15 @@ import { blurActiveElement } from '../../internal/closeActiveElement.js';
  *
  * @dependency syn-divider
  * @dependency syn-drawer
+ * @dependency syn-icon
+ * @dependency syn-nav-item
  *
  * @slot - The main content of the side-nav. Used for <syn-nav-item /> elements.
  * @slot footer - The footer content of the side-nav. Used for <syn-nav-item /> elements.
  *    Please avoid having to many nav-items as it can massively influence the user experience.
+ * @slot toggle-label - The label of the toggle nav-item for variant="sticky".
+ * @slot toggle-icon - An icon to use in lieu of the default icon for the toggle nav-item
+ * for variant="sticky".
  *
  * @event syn-show - Emitted when the side-nav opens.
  * @event syn-after-show - Emitted after the side-nav opens and all animations are complete.
@@ -54,22 +61,32 @@ import { blurActiveElement } from '../../internal/closeActiveElement.js';
  * @csspart panel - The side-nav's panel (where the whole content is rendered).
  * @csspart body - The side-nav's body (where the default slot content is rendered)
  * @csspart drawer__base - The drawer's base wrapper
- *
+ * @csspart toggle-nav-item - The nav-item to toggle open state for variant="sticky"
+ * @csspart toggle-icon - The icon of the toggle nav-item for variant="sticky"
+ * @csspart toggle-label - The label of the toggle nav-item for variant="sticky".
+
  * @cssproperty  --side-nav-open-width - The width of the side-nav if in open state
  *
- * @animation sideNav.showNonRail - The animation to use when showing the side-nav in non-rail mode.
- * @animation sideNav.showRail - The animation to use when showing the side-nav in rail mode.
- * @animation sideNav.hideNonRail - The animation to use when hiding the side-nav in non-rail mode.
- * @animation sideNav.hideRail - The animation to use when hiding the side-nav in rail mode.
+ * @animation sideNav.showNonRail - The animation to use when showing the side-nav
+ *  in variant="default".
+ * @animation sideNav.showRail - The animation to use when showing the side-nav in variant="rail"
+ *  and variant="sticky".
+ * @animation sideNav.hideNonRail - The animation to use when hiding the side-nav
+ *  in variant="default".
+ * @animation sideNav.hideRail - The animation to use when hiding the side-nav in variant="rail"
+ *  and variant="sticky".
  * @animation sideNav.overlay.show - The animation to use when showing the side-nav's overlay.
  * @animation sideNav.overlay.hide - The animation to use when hiding the side-nav's overlay.
  */
+@enableDefaultSettings('SynSideNav')
 export default class SynSideNav extends SynergyElement {
   static styles: CSSResultGroup = [componentStyles, styles];
 
   static dependencies = {
     'syn-divider': SynDivider,
     'syn-drawer': SynDrawer,
+    'syn-icon': SynIcon,
+    'syn-nav-item': SynNavItem,
   };
 
   private readonly hasSlotController = new HasSlotController(this, '[default]', 'footer');
@@ -93,13 +110,18 @@ export default class SynSideNav extends SynergyElement {
    * You can toggle this attribute to show and hide the side-nav, or you can use the `show()` and
    * `hide()` methods and this attribute will reflect the side-nav's open state.
    *
-   * Depending if the rail attribute is set or not, the behavior will differ.
+   * Depending on the "variant" attribute, the behavior will differ.
    *
-   * __Non rail__:
-   * With `open` will show the side-nav.
+   * __Default__:
+   * With `open` will show the side-nav with an overlay.
    * Without `open`, the side-nav will be hidden.
    *
    * __Rail__:
+   * With `open` will show the whole side-nav with an overlay for touch devices
+   * or without an overlay for non-touch devices.
+   * Without `open`, the side-nav will only show the prefix of nav-item's.
+   *
+   * __Sticky__:
    * With `open` will show the whole side-nav with an overlay for touch devices
    * or without an overlay for non-touch devices.
    * Without `open`, the side-nav will only show the prefix of nav-item's.
@@ -114,11 +136,34 @@ export default class SynSideNav extends SynergyElement {
    *
    * Note: The Rail is only an option if all Navigation Items on the first level have an Icon.
    * If this is not the case you should use a burger navigation.
+   *
+   * @deprecated Use the `variant` attribute with `rail` instead.
+   * Will be removed in synergy version 3.0
    */
   @property({ reflect: true, type: Boolean }) rail = false;
 
   /**
-   * By default, the side-nav traps the focus if in non-rail mode and open.
+   * The variant that should be used to show the side navigation.
+   *
+   * The following variants are supported:
+   * - **default** (default): Always shows the whole content and additionally an overlay.
+   * This makes especially sense for applications, where you navigate to a place and stay
+   * there for a longer time.
+   * - **rail**: Only show the prefix of navigation items in closed state.
+   * This will open on hover on the rail navigation.
+   * On touch devices the navigation opens on click and shows an overlay.
+   * Note: The rail variant is only an option if all Navigation Items on the first level
+   * have an Icon.
+   * If this is not the case you should use a burger navigation.
+   * - **sticky**: The side-nav has a pin button to show the side-nav in small (icon only)
+   * and full width. This variant is only possible for non-nested navigation items.
+   * Note: The sticky variant is only an option if all Navigation Items on the first level
+   * have an Icon and if there are only "first level" items.
+   */
+  @property({ reflect: true }) variant: 'default' | 'rail' | 'sticky' = 'default';
+
+  /**
+   * By default, the side-nav traps the focus if in variant="default" and open.
    * To disable the focus trapping, set this attribute.
    */
   @property({ attribute: 'no-focus-trapping', reflect: true, type: Boolean }) noFocusTrapping = false;
@@ -129,14 +174,14 @@ export default class SynSideNav extends SynergyElement {
   }
 
   private handleMouseEnter() {
-    // Debounce mouse events, to avoid infinite loop of open / closing in rail mode
+    // Debounce mouse events, to avoid infinite loop of open / closing in variant="rail"
     this.setDelayedCallback(() => {
       this.open = true;
     });
   }
 
   private handleMouseLeave() {
-    // Debounce mouse events, to avoid infinite loop of open / closing in rail mode
+    // Debounce mouse events, to avoid infinite loop of open / closing in variant="rail"
     this.setDelayedCallback(() => {
       this.open = false;
     });
@@ -158,21 +203,9 @@ export default class SynSideNav extends SynergyElement {
     this.drawer.shadowRoot!.querySelector('.drawer__panel')?.removeEventListener('mouseleave', this.handleMouseLeave);
   }
 
-  private setDrawerVisibility(isVisible: boolean) {
-    (this.drawer.shadowRoot!.querySelector('.drawer') as HTMLElement).hidden = !isVisible;
-    this.drawer.shadowRoot!.querySelector('.drawer__panel')?.setAttribute('aria-hidden', isVisible ? 'false' : 'true');
-  }
-
-  private forceDrawerVisibilityForRailMode() {
-    return waitForEvent(this.drawer, 'syn-after-hide').then(() => {
-      this.setDrawerVisibility(true);
-      this.isAnimationActive = false;
-    });
-  }
-
   private setDrawerAnimations() {
-    const showAnimation = getAnimation(this, `sideNav.show${this.rail ? 'Rail' : 'NonRail'}`, { dir: this.localize.dir() });
-    const hideAnimation = getAnimation(this, `sideNav.hide${this.rail ? 'Rail' : 'NonRail'}`, { dir: this.localize.dir() });
+    const showAnimation = getAnimation(this, `sideNav.show${this.variant === 'default' ? 'NonRail' : 'Rail'}`, { dir: this.localize.dir() });
+    const hideAnimation = getAnimation(this, `sideNav.hide${this.variant === 'default' ? 'NonRail' : 'Rail'}`, { dir: this.localize.dir() });
     const hideOverlay = getAnimation(this, 'sideNav.overlay.hide', { dir: this.localize.dir() });
     const showOverlay = getAnimation(this, 'sideNav.overlay.show', { dir: this.localize.dir() });
 
@@ -182,49 +215,40 @@ export default class SynSideNav extends SynergyElement {
     setAnimation(this.drawer, 'drawer.overlay.show', showOverlay);
   }
 
-  @watch('rail', { waitUntilFirstUpdate: true })
-  handleModeChange() {
+  @watch('variant', { waitUntilFirstUpdate: true })
+  handleVariantChange() {
     this.setDrawerAnimations();
+    this.drawer.forceVisibility(this.variant !== 'default');
 
-    if (this.rail) {
+    switch (this.variant) {
+    case 'rail':
+      // For hover handling
       this.addMouseListener();
-      // Force drawer visibility for rail mode
-      this.setDrawerVisibility(true);
-    } else {
+      break;
+    case 'sticky':
+    case 'default':
+    default:
       this.removeMouseListener();
-      // Remove forcing of drawer visibility for rail mode if not open
-      if (!this.open) {
-        this.setDrawerVisibility(false);
-      }
     }
   }
 
   @watch('open', { waitUntilFirstUpdate: true })
   handleOpenChange() {
-    if (!this.open) {
-      blurActiveElement(this);
-    }
-
-    if (!this.rail) {
+    if (this.variant === 'default') {
       return;
     }
 
     this.isAnimationActive = true;
 
-    if (!this.open) {
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.forceDrawerVisibilityForRailMode();
-    } else {
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      waitForEvent(this.drawer, 'syn-after-show').then(() => {
-        this.isAnimationActive = false;
-      });
-    }
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    waitForEvent(this.drawer, `syn-after-${this.open ? 'show' : 'hide'}`).then(() => {
+      this.isAnimationActive = false;
+    });
   }
 
   @watch('noFocusTrapping', { waitUntilFirstUpdate: true })
   handleFocusTrapping() {
-    if (!this.rail) {
+    if (this.variant === 'default') {
       if (this.noFocusTrapping) {
         this.drawer.modal.activateExternal();
       } else {
@@ -259,7 +283,7 @@ export default class SynSideNav extends SynergyElement {
     this.handleMouseEnter = this.handleMouseEnter.bind(this);
     this.handleMouseLeave = this.handleMouseLeave.bind(this);
     this.addEventListener('syn-initial-focus', (event) => {
-      if (this.rail) {
+      if (this.variant !== 'default') {
         // We need to do this, to stop the drawer from giving focus to the panel
         event.preventDefault();
 
@@ -272,8 +296,8 @@ export default class SynSideNav extends SynergyElement {
 
     this.addEventListener('focusin', (event) => {
       const targetTag = (event.target as HTMLElement).tagName.toLowerCase();
-      // Open the side-nav if it`s in rail mode, closed and the focused element is a nav-item
-      if (targetTag === 'syn-nav-item' && this.rail && !this.open) {
+      // Open the side-nav if it`s in variant="rail", closed and the focused element is a nav-item
+      if (targetTag === 'syn-nav-item' && this.variant === 'rail' && !this.open) {
         this.open = true;
       }
     });
@@ -282,38 +306,43 @@ export default class SynSideNav extends SynergyElement {
       const targetTag = (event.target as HTMLElement).tagName.toLowerCase();
       const relatedTargetTag = (event.relatedTarget as HTMLElement)?.tagName.toLowerCase();
 
-      // Close the side-nav, if it`s in rail mode, open and the next focused element
+      // Close the side-nav, if it`s in variant="rail", open and the next focused element
       // is no longer a nav-item
-      if (targetTag === 'syn-nav-item' && relatedTargetTag !== 'syn-nav-item' && this.rail && this.open) {
+      if (targetTag === 'syn-nav-item' && relatedTargetTag !== 'syn-nav-item' && this.variant === 'rail' && this.open) {
         this.open = false;
       }
     });
   }
 
   /**
-   * Initial setup for first render like special rail mode handling and drawer animations.
+   * Initial setup for first render like special variant="rail" and variant="sticky" handling
+   * and drawer animations.
    * */
   firstUpdated() {
     this.setDrawerAnimations();
 
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     this.drawer.updateComplete.then(() => {
+      this.drawer.forceVisibility(this.variant !== 'default');
       // change tabindex of drawer to make only nav-items focusable and not the panel of the drawer
       (this.drawer.shadowRoot!.querySelector('.drawer__panel') as HTMLElement).tabIndex = -1;
     });
 
-    if (this.rail) {
+    switch (this.variant) {
+    case 'rail':
       // Wait for the drawer`s update to be completed
-
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
       this.drawer.updateComplete.then(() => {
         this.addMouseListener();
-        // set initial visibility of drawer for rail mode
-        this.setDrawerVisibility(true);
       });
-    } else if (this.noFocusTrapping) {
-      // Disable the focus trapping of the modal
-      this.drawer.modal.activateExternal();
+      break;
+    case 'sticky': break;
+    case 'default':
+    default:
+      if (this.noFocusTrapping) {
+        // Disable the focus trapping of the modal
+        this.drawer.modal.activateExternal();
+      }
     }
   }
 
@@ -327,9 +356,36 @@ export default class SynSideNav extends SynergyElement {
     }
   }
 
+  // eslint-disable-next-line complexity
+  protected override willUpdate(changedProperties: PropertyValues) {
+    super.willUpdate(changedProperties);
+
+    // TODO: this can be removed in synergy version 3.0
+    if (changedProperties.has('rail')) {
+      if (this.rail) {
+        // Add deprecation console warning for stakeholder, which do not use linting
+        // to get their attention
+        console.warn('<syn-side-nav/>: The `rail` attribute is deprecated. Please use the `variant` attribute with `rail` instead. It will be removed in synergy version 3.0');
+      }
+
+      // The `variant` should be adapted to the `rail` attribute,
+      // if it was explicitly set or unset by the user.
+      // This is needed to be backwards compatible with the `rail` attribute
+      if (!changedProperties.has('variant') || this.rail) {
+        this.variant = this.rail ? 'rail' : 'default';
+      }
+    }
+  }
+
+  private toggleOpenState() {
+    this.open = !this.open;
+  }
+
+  // eslint-disable-next-line complexity
   render() {
     const isTouch = window.navigator.maxTouchPoints > 0 || !!('ontouchstart' in window);
     const hasFooter = this.hasSlotController.test('footer');
+    const showFooterDivider = hasFooter || this.variant === 'sticky';
 
     /* eslint-disable lit/no-invalid-html */
     /* eslint-disable @typescript-eslint/unbound-method */
@@ -338,10 +394,11 @@ export default class SynSideNav extends SynergyElement {
         class=${classMap({
           'side-nav': true,
           'side-nav--animation': this.isAnimationActive,
-          'side-nav--fix': !this.rail,
+          'side-nav--fix': this.variant === 'default',
           'side-nav--has-footer': hasFooter,
           'side-nav--open': this.open,
-          'side-nav--rail': this.rail,
+          'side-nav--rail': this.variant === 'rail',
+          'side-nav--sticky': this.variant === 'sticky',
           'side-nav--touch': isTouch,
         })}
         part="base"
@@ -349,7 +406,7 @@ export default class SynSideNav extends SynergyElement {
         
         <syn-drawer
           class="side-nav__drawer"
-          ?contained=${this.rail}
+          ?contained=${this.variant !== 'default'}
           exportparts="overlay,panel,body,base:drawer__base"
           label=${this.localize.term('sideNav')}
           no-header
@@ -364,8 +421,19 @@ export default class SynSideNav extends SynergyElement {
           
           <footer class="side-nav__footer" part="footer-container" slot="footer">  
 
-            ${hasFooter ? html`<syn-divider part="footer-divider" class="side-nav__footer-divider"></syn-divider>` : ''}
-            <slot name="footer" part="footer" ></slot> 
+            ${showFooterDivider ? html`<syn-divider part="footer-divider" class="side-nav__footer-divider"></syn-divider>` : ''}
+            <slot name="footer" part="footer" ></slot>
+            ${this.variant === 'sticky'
+              ? html`<syn-nav-item part="toggle-nav-item" class="side-nav__toggle-nav-item" @click=${this.toggleOpenState} ?divider=${hasFooter}>
+                      <slot name="toggle-icon" slot="prefix" class="side-nav__toggle-icon">
+                        <syn-icon library="system" name="sticky_sidebar" part="toggle-icon"></syn-icon>
+                      </slot>
+                      <slot name="toggle-label" part="toggle-label">
+                        ${(!this.open && !this.isAnimationActive) ? this.localize.term('sideNavShow') : this.localize.term('sideNavHide')}
+                      </slot>
+                    </syn-nav-item>`
+              : ''
+            }
           
           </footer>