+++ packages/components/CHANGELOG.md
@@ -1,3 +1,11 @@
+# [@synergy-design-system/components-v2.32.0](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.31.2...components/2.32.0) (2025-05-21)
+
+
+### Features
+
+* ✨ Add variant="sticky" to syn-side-nav ([#855](https://github.com/synergy-design-system/synergy-design-system/issues/855)) ([0004497](https://github.com/synergy-design-system/synergy-design-system/commit/0004497ff3c35fab1de65fdd70730f5962ffd748))
+ > `rail` attribute received a deprecation warning as it will be removed in the next major synergy version (3.0). Please use `variant="rail"` instead.
+
 # [@synergy-design-system/components-v2.31.2](https://github.com/synergy-design-system/synergy-design-system/compare/components/2.31.1...components/2.31.2) (2025-05-20)